requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.1
python-dotenv>=1.0.1
ultralytics>=8.2.0
opencv-python-headless>=4.9.0.80
numpy>=2.0.0
scipy>=1.13.0
Pillow>=10.3.0
shapely>=2.0.4
requests>=2.31.0
python-multipart>=0.0.9
